Rebrand DumpVisitor as PrettyPrintVisitor
