+++ javaparser-core/src/main/java/com/github/javaparser/printer/PrettyPrinter.java
@@ -1,0 +1,20 @@
+package com.github.javaparser.printer;
+
+import com.github.javaparser.ast.Node;
+
+/**
+ * Pretty printer for AST nodes.
+ */
+public class PrettyPrinter {
+    private final PrettyPrinterConfiguration configuration;
+
+    public PrettyPrinter(PrettyPrinterConfiguration configuration) {
+        this.configuration = configuration;
+    }
+
+    public String print(Node node) {
+        final PrettyPrintVisitor visitor = new PrettyPrintVisitor(configuration);
+        node.accept(visitor, null);
+        return visitor.getSource();
+    }
+}
